+++ lab/Lab3/src/Inst.h
@@ -18,28 +18,28 @@ class Inst
 public:
     Inst() {};
     ~Inst() {};
-    Inst(string name, double x, double y, double width, double height, bool fixed) : name(name), x(x), y(y), width(width), height(height), fixed(fixed) {};
+    Inst(string name, int x, int y, int width, int height, bool fixed) : name(name), x(x), y(y), width(width), height(height), fixed(fixed) {};
     Inst(const Inst &inst) : name(inst.name), x(inst.x), y(inst.y), width(inst.width), height(inst.height), fixed(inst.fixed) {};
 
-    void setX(double x) { this->x = x; }
-    void setY(double y) { this->y = y; }
-    void setWidth(double width) { this->width = width; }
-    void setHeight(double height) { this->height = height; }
+    void setX(int x) { this->x = x; }
+    void setY(int y) { this->y = y; }
+    void setWidth(int width) { this->width = width; }
+    void setHeight(int height) { this->height = height; }
     void setFixed(bool fixed) { this->fixed = fixed; }
 
     string getName() const { return name; }
-    double getX() const { return x; }
-    double getY() const { return y; }
-    double getWidth() const { return width; }
-    double getHeight() const { return height; }
+    int getX() const { return x; }
+    int getY() const { return y; }
+    int getWidth() const { return width; }
+    int getHeight() const { return height; }
     bool getFixed() const { return fixed; }
 
     friend ostream &operator<<(ostream &os, const Inst &inst);
 
 private:
     string name;
-    double x, y;
-    double width, height;
+    int x, y;
+    int width, height;
     bool fixed;
 };
